tools-lambda-function-construct/requirements.txt
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: Apache-2.0

aws-cdk-lib==2.232.1
constructs>=10.0.0,<11.0.0
